requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bitarray==2.8.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.6.1
Deprecated==1.2.14
dnspython==2.4.2
docx2txt==0.8
Flask==2.3.3
Flask-Limiter==3.5.0
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==2.0.2
idna==3.4
importlib-resources==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.305
langsmith==0.0.41
libmagic==1.0
limits==3.6.0
llama-index==0.8.36
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
ordered-set==4.1.0
packaging==23.1
pandas==2.1.1
pdfreader==0.1.12
Pillow==10.0.1
pluggy==1.3.0
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pylibmagic==0.4.0
pymongo==4.5.0
pypdf==3.16.2
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==7.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
